+++ NerdyWeirdWords/app/routes.ts
@@ -1,4 +1,11 @@
-import { type RouteConfig, index } from "@react-router/dev/routes";
+import {
+    type RouteConfig,
+    index,
+    route,
+} from "@react-router/dev/routes";
 
-export default [index("routes/home.tsx")] satisfies RouteConfig;
+export default [
+    index("routes/home.tsx"),
+    route("signin", "user/signInForm.tsx"),
+] satisfies RouteConfig;
